+++ packages/control/algorithm/surplus_controlled_test.py
@@ -4,7 +4,8 @@ import pytest
 
 from control.algorithm import surplus_controlled
 from control.algorithm.surplus_controlled import SurplusControlled
-from control.chargepoint.chargepoint import Chargepoint, ChargepointData, Get, Set
+from control.chargepoint.chargepoint import Chargepoint, ChargepointData
+from control.chargepoint.chargepoint_data import Get, Set
 from control.chargepoint.control_parameter import ControlParameter
 from control.ev import ChargeTemplate, Ev
